Fixed compiler warning

+++ src/worldmap/manager.cpp
@@ -321,7 +321,7 @@ WorldMapComponent::draw (DrawingContext& gc)
 
   scene_context->clear();
   scene_context->push_modelview();
-  scene_context->translate(cliprect.left, cliprect.top);
+  scene_context->translate((float)cliprect.left, (float)cliprect.top);
 
   scene_context->set_cliprect(cliprect);
